validaciones

+++ src/router/rutaCarrito.js
@@ -8,7 +8,12 @@ const data = new Contenedor('./data/pedidos.txt')
 rutaCarrito.get('/:id/productos',async (req,res)=>{
     const {id} = req.params
     const carrito = await data.getById(id)
-    res.send(carrito)
+    if(carrito){
+            return res.send(carrito)
+    } else{
+        res.status(404).send({ message: 'Error el carrito no existe' })
+    }
+    
 })
 
 //Crea nuevo carrito
@@ -20,22 +25,44 @@ rutaCarrito.post('/', async(req,res)=>{
 //Elimina un carrito
 rutaCarrito.delete('/:id', async(req,res)=>{
     const {id}= req.params
-    const carrito = await data.deleteById(id)
-    res.send(carrito)
+    const existe = await data.getById(id)
+
+    if(!existe){
+        return res.status(404).send({ message: 'Error el carrito no existe' })
+    }else{
+        const carrito = await data.deleteById(id)
+        return res.send(carrito)
+        
+    }
+    
 })
 
 //Agrega un producto al carrito
 rutaCarrito.post('/:id/productos', async(req,res)=>{
     const {id} = req.params
     const modificacion = req.body
-    const carrito = await data.moreProd(id, modificacion)
-    res.send(carrito)
+    const existe = await data.getById(id)
+
+    if(!existe){
+        return res.status(404).send({ message: 'Error el carrito no existe' })
+    }else{
+        const carrito = await data.moreProd(id, modificacion)
+        res.send(carrito)
+        
+    }
+    
 })
 
 //Elimina un producto del carrito
 rutaCarrito.delete('/:id/productos/:id_prod', async(req,res)=>{
     const {id, id_prod} = req.params
-    const carrito = await data.deleteOneProd(id, id_prod)
-    res.send(carrito)
+    const existe = await data.getById(id)
+
+    if(!existe){
+        return res.status(404).send({ message: 'Error el carrito no existe' })
+    }else{
+        const carrito = await data.deleteOneProd(id, id_prod)
+        res.send(carrito)
+    }
 })
 module.exports= rutaCarrito
